+++ Source/RendererRuntime/Public/Resource/MaterialBlueprint/BufferManager/LightBufferManager.h
@@ -134,13 +134,13 @@ namespace RendererRuntime
 		*    Fill the light buffer
 		*
 		*  @param[in] worldSpaceCameraPosition
-		*    World space camera position for camera relative rendering
+		*    64 bit world space position of the camera for camera relative rendering
 		*  @param[in] sceneResource
 		*    Scene resource to use
 		*  @param[out] commandBuffer
 		*    Command buffer to fill
 		*/
-		void fillBuffer(const glm::vec3& worldSpaceCameraPosition, SceneResource& sceneResource, Renderer::CommandBuffer& commandBuffer);
+		void fillBuffer(const glm::dvec3& worldSpaceCameraPosition, SceneResource& sceneResource, Renderer::CommandBuffer& commandBuffer);
 
 		/**
 		*  @brief
@@ -189,7 +189,7 @@ namespace RendererRuntime
 	private:
 		explicit LightBufferManager(const LightBufferManager&) = delete;
 		LightBufferManager& operator=(const LightBufferManager&) = delete;
-		void fillTextureBuffer(const glm::vec3& worldSpaceCameraPosition, SceneResource& sceneResource);
+		void fillTextureBuffer(const glm::dvec3& worldSpaceCameraPosition, SceneResource& sceneResource);	// 64 bit world space position of the camera
 		void fillClusters3DTexture(SceneResource& sceneResource, Renderer::CommandBuffer& commandBuffer);
